refactor: move `normalizeFunctionsConfig` to a separate file
