+++ src/commands/deploy.js
@@ -8,10 +8,10 @@ const inquirer = require('inquirer')
 const isObject = require('lodash/isObject')
 const prettyjson = require('prettyjson')
 
-const { normalizeFunctionsConfig } = require('../function-builder-detectors/zisi')
 const { cancelDeploy } = require('../lib/api')
 const { getBuildOptions, runBuild } = require('../lib/build')
 const { statAsync } = require('../lib/fs')
+const { normalizeFunctionsConfig } = require('../lib/functions/config')
 const { getLogMessage } = require('../lib/log')
 const { startSpinner, stopSpinner } = require('../lib/spinner')
 const Command = require('../utils/command')
